Bluetooth: Update OPP progress bar with right time stamp
CRs-fixed: 364121

+++ src/com/android/bluetooth/opp/BluetoothOppNotification.java
@@ -1,5 +1,6 @@
 /*
  * Copyright (c) 2008-2009, Motorola, Inc.
+ * Copyright (c) 2012, Code Aurora Forum. All rights reserved.
  *
  * All rights reserved.
  *
@@ -111,7 +112,7 @@ class BluetoothOppNotification {
 
         int totalTotal = 0; // total bytes for current transfer
 
-        int timeStamp = 0; // Database time stamp. Used for sorting ongoing transfers.
+        long timeStamp = 0; // Database time stamp. Used for sorting ongoing transfers.
 
         String description; // the text above progress bar
     }
@@ -228,7 +229,7 @@ class BluetoothOppNotification {
 
         mNotifications.clear();
         for (cursor.moveToFirst(); !cursor.isAfterLast(); cursor.moveToNext()) {
-            int timeStamp = cursor.getInt(timestampIndex);
+            long timeStamp = cursor.getLong(timestampIndex);
             int dir = cursor.getInt(directionIndex);
             int id = cursor.getInt(idIndex);
             int total = cursor.getInt(totalBytesIndex);
